feat: Add black code formatter
This commit adds the black code formatter to the project. Black is an uncompromising Python code formatter that provides speed, determinism, and freedom from pycodestyle nagging about formatting. It reformats entire files in place and produces the smallest diffs possible, making code review faster. The black version added is 23.10.1.

For more information on how to install and use Black, please refer to the documentation.

Co-authored-by: Łukasz Langa <lukasz@langa.pl>